+++ zcash_client_sqlite/src/testing/pool.rs
@@ -52,7 +52,10 @@ use zcash_protocol::consensus::BlockHeight;
 use super::TestFvk;
 use crate::{
     error::SqliteClientError,
-    testing::{input_selector, AddressType, BlockCache, InitialChainState, TestBuilder, TestState},
+    testing::{
+        input_selector, AddressType, BlockCache, FakeCompactOutput, InitialChainState, TestBuilder,
+        TestState,
+    },
     wallet::{block_max_scanned, commitment_tree, parse_scope, truncate_to_height},
     AccountId, NoteId, ReceivedNoteId,
 };
@@ -1373,9 +1376,11 @@ pub(crate) fn checkpoint_gaps<T: ShieldedPoolTester>() {
     st.generate_block_at(
         account.birthday().height() + 10,
         BlockHash([0; 32]),
-        &not_our_key,
-        AddressType::DefaultExternal,
-        not_our_value,
+        &[FakeCompactOutput::new(
+            &not_our_key,
+            AddressType::DefaultExternal,
+            not_our_value,
+        )],
         st.latest_cached_block().unwrap().sapling_end_size,
         st.latest_cached_block().unwrap().orchard_end_size,
         false,
@@ -1955,9 +1960,11 @@ pub(crate) fn invalid_chain_cache_disconnected<T: ShieldedPoolTester>() {
     st.generate_block_at(
         disconnect_height,
         BlockHash([1; 32]),
-        &dfvk,
-        AddressType::DefaultExternal,
-        NonNegativeAmount::const_from_u64(8),
+        &[FakeCompactOutput::new(
+            &dfvk,
+            AddressType::DefaultExternal,
+            NonNegativeAmount::const_from_u64(8),
+        )],
         2,
         2,
         true,