+++ zcash_client_sqlite/CHANGELOG.md
@@ -8,8 +8,10 @@ and this library adheres to Rust's notion of
 ## [Unreleased]
 ### Changed
 - MSRV is now 1.70.0.
-- `SqliteClientError` enum
-  - Added `AccountCollision` variant.
+- `zcash_client_sqlite::error::SqliteClientError` has changed variants:
+  - Removed `HdwalletError`.
+  - Added `AccountCollision`.
+  - Added `TransparentDerivation`.
 
 ## [0.10.3] - 2024-04-08